Safari-611.1.10/LayoutTests/loader/navigation-policy/should-open-external-urls/main-frame-click-expected.txt

 - decidePolicyForNavigationAction
<NSURLRequest URL resources/click-notify-done-in-main.html, main document URL resources/click-notify-done-in-main.html, http method GET> is main frame - yes should open URLs externally - no
 - decidePolicyForNavigationAction
<NSURLRequest URL notify-done.html, main document URL notify-done.html, http method GET> is main frame - yes should open URLs externally - yes
